[refactor/WF-120]: refactor absolute paths
