+++ src/auth/auth.service.ts
@@ -1,13 +1,13 @@
-import { PositionsService } from './../positions/positions.service';
-import { CompaniesService } from './../companies/companies.service';
+import { PositionsService } from 'positions/positions.service';
+import { CompaniesService } from 'companies/companies.service';
 import { UserReturnDto } from './dto';
-import { TokensService } from './../tokens/tokens.service';
+import { TokensService } from 'tokens/tokens.service';
 import { LoginUserDto } from './dto/login-user.dto';
-import { UsersService } from './../users/users.service';
+import { UsersService } from 'users/users.service';
 import { Injectable } from '@nestjs/common';
 import { ForbiddenException, NotFoundException, UnauthorizedException } from '@nestjs/common/exceptions';
 import { compare } from 'bcrypt';
-import { PureUserDto, PureRelationsUserDto } from 'src/users/dto';
+import { PureUserDto, PureRelationsUserDto } from 'users/dto';
 import { RegisterUserOwnerDto, SetSessionReturnDto, UserSessionDto } from './dto';
 import { EMAIL_PASSWORD_INCORRECT, NOT_FOUND, NO_SESSION, USER_EXISTS } from '@constants/error';
 import { User } from '@prisma/client';
