+++ src/auth/dto/login-user.dto.ts
@@ -1,3 +1,3 @@
 import { PickType } from '@nestjs/swagger';
-import { CreateUserDto } from 'src/users/dto';
+import { CreateUserDto } from 'users/dto';
 export class LoginUserDto extends PickType(CreateUserDto, ['email', 'password'] as const) {}
