+++ src/auth/dto/register-user-owner.dto.ts
@@ -1,5 +1,5 @@
-import { CreateUserDto } from 'src/users/dto';
-import { CreateCompanyDto } from 'src/companies/dto';
+import { CreateUserDto } from 'users/dto';
+import { CreateCompanyDto } from 'companies/dto';
 
 export class RegisterUserOwnerDto {
   company: CreateCompanyDto;
