+++ src/auth/dto/user-session.dto.ts
@@ -1,5 +1,5 @@
 import { PickType } from '@nestjs/swagger';
-import { CreateUserDto } from 'src/users/dto';
+import { CreateUserDto } from 'users/dto';
 
 export class UserSessionDto extends PickType(CreateUserDto, [
   'password',
