+++ src/colors/colors.service.ts
@@ -1,4 +1,4 @@
-import { PrismaService } from './../prisma/prisma.service';
+import { PrismaService } from 'prisma/prisma.service';
 import { Injectable } from '@nestjs/common';
 import { FindByValue } from './interfaces';
