+++ src/common/decorators/public.decorator.ts
@@ -1,4 +1,4 @@
 import { SetMetadata } from '@nestjs/common';
-import { IS_PUBLIC_KEY } from '../constants';
+import { IS_PUBLIC_KEY } from '@constants/index';
 
 export const Public = () => SetMetadata(IS_PUBLIC_KEY, true);
