+++ src/common/guards/csrfAndAccessToken.guard.ts
@@ -1,7 +1,7 @@
-import { TokensService } from '../../tokens/tokens.service';
+import { TokensService } from 'tokens/tokens.service';
 import { CanActivate, ExecutionContext, Injectable, UnauthorizedException } from '@nestjs/common';
 import { Reflector } from '@nestjs/core';
-import { IS_PUBLIC_KEY } from '../constants';
+import { IS_PUBLIC_KEY } from '@constants/index';
 
 @Injectable()
 export class CsrfAndAccessTokenGuard implements CanActivate {
