+++ src/common/guards/refreshToken.guard.ts
@@ -1,5 +1,5 @@
 import { CanActivate, ExecutionContext, Injectable, UnauthorizedException } from '@nestjs/common';
-import { TokensService } from 'src/tokens/tokens.service';
+import { TokensService } from 'tokens/tokens.service';
 
 @Injectable()
 export class RefreshTokenGuard implements CanActivate {
