+++ src/companies/companies.module.ts
@@ -1,5 +1,5 @@
-import { TokensModule } from './../tokens/tokens.module';
-import { PrismaModule } from './../prisma/prisma.module';
+import { TokensModule } from 'tokens/tokens.module';
+import { PrismaModule } from 'prisma/prisma.module';
 import { Module } from '@nestjs/common';
 import { CompaniesService } from './companies.service';
 import { CompaniesController } from './companies.controller';
