+++ src/companies/dto/company-data.dto.ts
@@ -1,6 +1,6 @@
 import { ApiProperty } from '@nestjs/swagger';
 import { IsEmail, IsNotEmpty, IsString } from 'class-validator';
-import { IsPhoneNumber } from 'src/common/decorators';
+import { IsPhoneNumber } from '@decorators/index';
 
 export class CompanyDataDto {
   @ApiProperty({ example: '1efe537f-e380-4168-959b-f864f2196369' })
