+++ src/folders/appearance/folder-appearances.module.ts
@@ -1,10 +1,10 @@
 import { FolderAppearancesController } from './folder-appearances.controller';
 import { FoldersModule } from '../folders.module';
-import { TokensModule } from '../../tokens/tokens.module';
-import { ColorsModule } from '../../colors/colors.module';
+import { TokensModule } from 'tokens/tokens.module';
+import { ColorsModule } from 'colors/colors.module';
 import { FolderIconsService } from './services/folder-icons.service';
 import { FolderAppearancesService } from './folder-appearances.service';
-import { PrismaModule } from '../../prisma/prisma.module';
+import { PrismaModule } from 'prisma/prisma.module';
 import { Module } from '@nestjs/common';
 
 @Module({
