+++ src/folders/appearance/folder-appearances.service.ts
@@ -1,8 +1,8 @@
 import { FoldersService } from '../folders.service';
-import { TokensService } from '../../tokens/tokens.service';
+import { TokensService } from 'tokens/tokens.service';
 import { FolderIconsService } from './services/folder-icons.service';
-import { ColorsService } from '../../colors/colors.service';
-import { PrismaService } from '../../prisma/prisma.service';
+import { ColorsService } from 'colors/colors.service';
+import { PrismaService } from 'prisma/prisma.service';
 import { BadRequestException, ForbiddenException, Injectable, NotFoundException } from '@nestjs/common';
 import { NOTHING_PASSED, NOT_FOUND } from '@constants/error';
 import {
