+++ src/folders/appearance/services/folder-icons.service.ts
@@ -1,4 +1,4 @@
-import { PrismaService } from '../../../prisma/prisma.service';
+import { PrismaService } from 'prisma/prisma.service';
 import { Injectable } from '@nestjs/common';
 
 @Injectable()
