+++ src/folders/dto/user-folder.dto.ts
@@ -1,4 +1,4 @@
-import { PureUserDto } from './../../users/dto/pure-user.dto';
+import { PureUserDto } from 'users/dto/pure-user.dto';
 import { PickType } from '@nestjs/swagger';
 
 export class UserFolderDto extends PickType(PureUserDto, ['id', 'fullName', 'avatar'] as const) {}
