+++ src/folders/folders.module.ts
@@ -1,9 +1,9 @@
-import { TokensModule } from './../tokens/tokens.module';
-import { PrismaModule } from './../prisma/prisma.module';
+import { TokensModule } from 'tokens/tokens.module';
+import { PrismaModule } from 'prisma/prisma.module';
 import { forwardRef, Module } from '@nestjs/common';
 import { FoldersController } from './folders.controller';
 import { FoldersService } from './folders.service';
-import { UsersModule } from 'src/users/users.module';
+import { UsersModule } from 'users/users.module';
 import { FolderAppearancesModule } from './appearance/folder-appearances.module';
 import { FolderTypesService } from './services/folder-types.service';
