+++ src/folders/services/folder-types.service.ts
@@ -1,4 +1,4 @@
-import { PrismaService } from '../../prisma/prisma.service';
+import { PrismaService } from 'prisma/prisma.service';
 import { Injectable, NotFoundException } from '@nestjs/common';
 import { NOT_FOUND } from '@constants/error';
