+++ src/positions/positions.module.ts
@@ -1,4 +1,4 @@
-import { PrismaModule } from './../prisma/prisma.module';
+import { PrismaModule } from 'prisma/prisma.module';
 import { Module } from '@nestjs/common';
 import { PositionsService } from './positions.service';
