+++ src/prisma/prisma.service.ts
@@ -7,6 +7,10 @@ export class PrismaService extends PrismaClient implements OnModuleInit {
     await this.$connect();
   }
 
+  async onModuleDestroy() {    
+    await this.$disconnect();
+  }
+
   async enableShutdownHooks(app: INestApplication) {
     this.$on('beforeExit', async () => {
       await app.close();
