+++ src/tokens/tokens.module.ts
@@ -1,4 +1,4 @@
-import { PrismaModule } from './../prisma/prisma.module';
+import { PrismaModule } from 'prisma/prisma.module';
 import { Module } from '@nestjs/common';
 import { TokensService } from './tokens.service';
 import { JwtModule } from '@nestjs/jwt';
