+++ src/users/dto/find-user.dto.ts
@@ -1,4 +1,4 @@
-import { PositionDataDto } from 'src/positions/dto';
+import { PositionDataDto } from 'positions/dto';
 import { PureUserDto } from './pure-user.dto';
 
 export class FindUserDto {
