+++ src/users/dto/pure-user.dto.ts
@@ -1,4 +1,4 @@
-import { IsPhoneNumber } from './../../common/decorators';
+import { IsPhoneNumber } from '@decorators/index';
 import { ApiProperty } from '@nestjs/swagger';
 import { User } from '@prisma/client';
 import { IsDateString, IsEmail, IsNotEmpty, IsString } from 'class-validator';
