+++ tsconfig.json
@@ -9,7 +9,7 @@
     "target": "es2017",
     "sourceMap": true,
     "outDir": "./dist",
-    "baseUrl": "./",
+    "baseUrl": "./src",
     "incremental": true,
     "skipLibCheck": true,
     "strictNullChecks": false,
@@ -18,7 +18,9 @@
     "forceConsistentCasingInFileNames": false,
     "noFallthroughCasesInSwitch": false,
     "paths": {
-      "@constants/*": ["./src/common/constants/*"]
+      "@constants/*": ["common/constants/*"],
+      "@decorators/*": ["common/decorators/*"],
+      "@guards/*": ["common/guards/*"]
     }
   }
 }
